--- mifosng-android/src/main/java/com/mifos/api/services/CreateSavingsAccountService.java
@@ -1,33 +1,0 @@
-/*
- * This project is licensed under the open source MPL V2.
- * See https://github.com/openMF/android-client/blob/master/LICENSE.md
- */
-package com.mifos.api.services;
-
-import com.mifos.api.model.APIEndPoint;
-import com.mifos.objects.client.Savings;
-import com.mifos.objects.organisation.ProductSavings;
-import com.mifos.objects.templates.savings.SavingProductsTemplate;
-import com.mifos.services.data.SavingsPayload;
-
-import java.util.List;
-
-import retrofit2.http.Body;
-import retrofit2.http.GET;
-import retrofit2.http.POST;
-import rx.Observable;
-
-/**
- * @author
- */
-public interface CreateSavingsAccountService {
-
-    @GET(APIEndPoint.CREATESAVINGSPRODUCTS)
-    Observable<List<ProductSavings>> getAllSavingsAccounts();
-
-    @POST(APIEndPoint.CREATESAVINGSACCOUNTS)
-    Observable<Savings> createSavingsAccount(@Body SavingsPayload savingsPayload);
-
-    @GET(APIEndPoint.CREATESAVINGSPRODUCTS + "/template")
-    Observable<SavingProductsTemplate> getSavingsAccountTemplate();
-}
